projects/chapter-19-project/requirements.txt
# Chapter 19 Project Requirements

# Web framework
fastapi>=0.104.0
uvicorn>=0.24.0

# Core AI library
openai>=1.0.0

# Database (choose based on needs)
# SQLAlchemy for SQL
sqlalchemy>=2.0.0
# pymongo for MongoDB (NoSQL)
# pymongo>=4.6.0
# Vector database (optional)
# chromadb>=0.4.0

# Message queue (optional)
# redis>=5.0.0

# Environment variable management
python-dotenv>=1.0.0
